CCMSPUI-786 | Renaming of package names from *.cwa.bulkupload.* to *.bulkclaim.*
