+++ settings.gradle
@@ -13,21 +13,6 @@ pluginManagement {
     }
 }
 
-pluginManagement {
-    repositories {
-        maven {
-            name = "gitHubPackages"
-            url = 'https://maven.pkg.github.com/ministryofjustice/laa-ccms-spring-boot-common'
-            credentials {
-                username = System.getenv("GITHUB_ACTOR")?.trim() ?: settings.ext.find('project.ext.gitPackageUser')
-                password = System.getenv("GITHUB_TOKEN")?.trim() ?: settings.ext.find('project.ext.gitPackageKey')
-            }
-        }
-        maven { url = "https://plugins.gradle.org/m2/" }
-        gradlePluginPortal()
-    }
-}
-
 plugins {
     // Apply the foojay-resolver plugin to allow automatic download of JDKs
     id 'org.gradle.toolchains.foojay-resolver-convention' version '1.0.0'
