+++ src/main/java/uk/gov/justice/laa/bulkclaim/config/rest/ApiProperties.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.cwa.bulkupload.config.rest;
+package uk.gov.justice.laa.bulkclaim.config.rest;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
