+++ src/main/java/uk/gov/justice/laa/bulkclaim/config/rest/WebClientConfiguration.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.cwa.bulkupload.config.rest;
+package uk.gov.justice.laa.bulkclaim.config.rest;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -10,7 +10,7 @@ import org.springframework.web.reactive.function.client.ExchangeStrategies;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.support.WebClientAdapter;
 import org.springframework.web.service.invoker.HttpServiceProxyFactory;
-import uk.gov.justice.laa.cwa.bulkupload.service.ClaimsRestService;
+import uk.gov.justice.laa.bulkclaim.service.ClaimsRestService;
 
 /**
  * Configuration class for creating and configuring WebClient instances.
