+++ src/main/java/uk/gov/justice/laa/bulkclaim/controller/HomeController.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.cwa.bulkupload.controller;
+package uk.gov.justice.laa.bulkclaim.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
