+++ src/main/java/uk/gov/justice/laa/bulkclaim/validation/BulkImportFileVirusValidator.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.cwa.bulkupload.validation;
+package uk.gov.justice.laa.bulkclaim.validation;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
 import org.springframework.web.multipart.MultipartFile;
-import uk.gov.justice.laa.cwa.bulkupload.dto.FileUploadForm;
-import uk.gov.justice.laa.cwa.bulkupload.service.VirusCheckService;
+import uk.gov.justice.laa.bulkclaim.dto.FileUploadForm;
+import uk.gov.justice.laa.bulkclaim.service.VirusCheckService;
 
 /**
  * A validator for validating files intended for bulk submissions. This class implements the {@link
