+++ src/main/resources/templates/pages/upload.html
@@ -7,7 +7,7 @@
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds-from-desktop">
 
-      <!--/*@thymesVar id="fileUploadForm" type="uk.gov.justice.laa.cwa.bulkupload.dto.FileUploadForm"*/-->
+      <!--/*@thymesVar id="fileUploadForm" type="uk.gov.justice.laa.bulkclaim.dtoadForm"*/-->
       <form class="form" th:action="@{/upload}" method="post"
             enctype="multipart/form-data" th:object="${fileUploadForm}">
