+++ src/test/java/uk/gov/justice/laa/bulkclaim/config/WebMvcTestConfig.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.cwa.bulkupload.config;
+package uk.gov.justice.laa.bulkclaim.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
