+++ wiremock/mappings/cwa-service/providerslist-200.json
@@ -4,7 +4,7 @@
     "urlPath": "/api/validate_user",
     "queryParameters": {
       "username": {
-        "matches": "DT_SCRIPT_USER4|DT_SCRIPT_USER14|DT_SCRIPT_USER19"
+        "matches": "DT_SCRIPT_USER4|DT_SCRIPT_USER14|DT_SCRIPT_USER19|provider.user@provider.com"
       }
     }
   },
@@ -14,14 +14,14 @@
     },
     "status": 200,
     "jsonBody": [
-                {
-                  "VENDOR_NAME": "ABC Solicitors",
-                  "VENDOR_ID": 1
-                },
-                {
-                  "VENDOR_NAME": "XYZ Solicitors",
-                  "VENDOR_ID": 2
-                }
+      {
+        "VENDOR_NAME": "ABC Solicitors",
+        "VENDOR_ID": 1
+      },
+      {
+        "VENDOR_NAME": "XYZ Solicitors",
+        "VENDOR_ID": 2
+      }
     ]
   }
 }
